sistema-extintores/requirements.txt
# Framework Web
Flask==3.0.0
Werkzeug==3.0.1

# Base de Datos
Flask-SQLAlchemy==3.1.1

# Comunicación en Tiempo Real
Flask-SocketIO==5.3.5
python-socketio==5.10.0
python-engineio==4.8.0

# CORS (Cross-Origin Resource Sharing)
Flask-CORS==4.0.0

# Generación de PDFs
reportlab==4.0.7

# Utilidades
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
psutil==5.9.6

# SSL/HTTPS
pyOpenSSL==24.0.0

# Servidor WSGI para producción
gunicorn==21.2.0
eventlet==0.33.3
